Put TO mode in a different class and make it actually look at the mode on init
